Grabpast and spotifyId updated
Moved `getSpotifyID` to config tools to expand its scope to other files. This discourages a circular import.
